Rename lookups.py to color_mapping.py
