+++ satpy/enhancements/color_mapping.py
